# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  controllers.NewCategory:
    properties:
      categoryName:
        example: Anime
        minLength: 2
        type: string
    required:
    - categoryName
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: API Server for Ozinwe
  title: Ozinwe API
  version: "1.0"
paths:
  /admin/category/{id}/delete:
    delete:
      consumes:
      - application/json
      operationId: delete-category
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: DeleteCategory
      tags:
      - admin-category-controller
  /admin/category/{id}/edit:
    get:
      consumes:
      - application/json
      operationId: edit-category
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: EditCategory
      tags:
      - admin-category-controller
  /admin/category/{id}/update:
    put:
      consumes:
      - application/json
      operationId: update-category
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: categoryName
        in: body
        name: categoryName
        required: true
        schema:
          $ref: '#/definitions/controllers.NewCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: UpdateCategory
      tags:
      - admin-category-controller
  /admin/category/create:
    post:
      consumes:
      - application/json
      operationId: create-category
      parameters:
      - description: categoryName
        in: body
        name: categoryName
        required: true
        schema:
          $ref: '#/definitions/controllers.NewCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: CreateCategory
      tags:
      - admin-category-controller
...
  /movie/create:
    post:
      consumes:
      - application/json
      operationId: create-movie
      parameters:
      - description: movieID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: CreateMovie
      tags:
      - admin-movie-controller
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"